Fixed dot product.

+++ AlgorithmicToolbox.Week03/src/DotProduct.java
@@ -35,7 +35,7 @@ public class DotProduct {
     		Arrays.sort(b);
         long result = 0;
         for (int i = 0; i < a.length; i++) {
-            result += a[i] * b[i];
+            result += ((long) a[i]) * ((long) b[i]);
         }
         return result;
     }
